[refactor] : ちょっとした名前の修正
