+++ src/utils/icons/index.ts
@@ -1,3 +1,3 @@
-export { default as SettingIcon } from './setting';
+export { default as SettingsIcon } from './settings';
 export { default as TaskIcon } from './task';
 export { default as QuestionIcon } from './question';
